file restructuring: moved styles, services, types, auth.ts, utils outside of app directory, configured correct import paths
