+++ frontend/src/app/api/auth/[...nextauth]/route.ts
@@ -1,4 +1,4 @@
-import { handlers } from '@/app/auth';
+import { handlers } from '@/auth';
