+++ frontend/src/auth.ts
@@ -1,6 +1,6 @@
 import NextAuth from "next-auth"
 import Google from "next-auth/providers/google"
-import userService from "@/app/services/user.service"
+import userService from "@/services/user.service"
 
 
 type SignInParams = {
