+++ frontend/src/components/MainPage/Card.tsx
