+++ frontend/src/components/MainPage/MainPageContent.tsx
