+++ frontend/src/style/DisplayData.css
